更改为greenDao,完成动态添加条目的功能

+++ app/src/main/java/com/gersion/superlock/utils/ConfigManager.java
@@ -32,7 +32,7 @@ import static com.gersion.superlock.utils.MyConstants.USER_NAME;
  */
 
 public class ConfigManager {
-    private int mDataListCount;
+    private long mDataListCount;
     private String mBackupEmailAddress;
     private boolean mSuperPasswordSeted;
     private int mLockType;
@@ -257,7 +257,7 @@ public class ConfigManager {
         SpfUtils.putString(mContext, BACKUP_EMAIL_ADDRESS,backupEmailAddress);
     }
 
-    public int getDataListCount() {
+    public long getDataListCount() {
         return mDataListCount;
     }
